database pull example and beginnings of the shiny app
databsePull.R shows how to connect to and run a query on our pension database. app.R is the start of the shiny app. As a first step it creates a drop down menu based on the states that we have plans for. Next it will allow selection of the specific plan and then create the output.